Buscar endereços finalizado

+++ backend/src/main/java/edu/senac/backend/cliente/EnderecosClienteRecord.java
@@ -2,7 +2,7 @@ package edu.senac.backend.cliente;
 
 public record EnderecosClienteRecord(
         Long id,
-        Integer idCliente,
+        Long idCliente,
         String cep,
         String logradouro,
         String numero,
